Pull out components, add stories for docs, hide in chromatic
