+++ src/components/table-of-contents/TableOfContentsItems.js
@@ -1,7 +1,10 @@
 import React, { useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { breakpoint, color, typography } from '../shared/styles';
+import { BulletLink } from './BulletLink';
+import { ItemLink } from './ItemLink';
+import { MenuLink } from './MenuLink';
+import { color, typography } from '../shared/styles';
 import { Icon } from '../Icon';
 import { Link } from '../Link';
 
@@ -27,100 +30,8 @@ const getItemComponent = (itemType) => {
   }
 };
 
-const StyledBulletLink = styled(({ isActive, ...rest }) => <Link {...rest} />)`
-  display: inline-block;
-  padding: 6px 0;
-  line-height: 1.5;
-  position: relative;
-  z-index: 1;
-  ${(props) => props.isActive && `font-weight: ${typography.weight.bold};`}
-
-  &::after {
-    position: absolute;
-    top: 0px;
-    right: auto;
-    bottom: 0px;
-    left: 3px;
-    width: auto;
-    height: auto;
-    border-left: 1px solid ${color.mediumlight};
-    content: '';
-    z-index: -1;
-  }
-`;
-
-const BulletLinkWrapper = styled.li`
-  && {
-    padding-top: 0;
-  }
-
-  &:first-of-type ${StyledBulletLink} {
-    margin-top: -6px;
-
-    &::after {
-      height: 50%;
-      top: 50%;
-    }
-  }
-
-  &:last-of-type ${StyledBulletLink} {
-    margin-bottom: -6px;
-
-    &::after {
-      height: 50%;
-      bottom: 50%;
-    }
-  }
-`;
-
-const Bullet = styled.span`
-  display: inline-block;
-  margin-bottom: 1px;
-  margin-right: 16px;
-  background: ${color.medium};
-  box-shadow: white 0 0 0 4px;
-  height: 8px;
-  width: 8px;
-  border-radius: 1em;
-  text-decoration: none !important;
-  content: '';
-  ${(props) => props.isActive && `background: ${color.secondary};`}
-`;
-
-function BulletLink({ currentPath, item, ...rest }) {
-  const isActive = currentPath === item.path;
-
-  return (
-    <BulletLinkWrapper>
-      <StyledBulletLink
-        isActive={isActive}
-        href={item.path}
-        LinkWrapper={item.LinkWrapper}
-        tertiary={!isActive}
-      >
-        <Bullet isActive={isActive} />
-        {item.title}
-      </StyledBulletLink>
-    </BulletLinkWrapper>
-  );
-}
-
-BulletLink.propTypes = {
-  currentPath: PropTypes.string.isRequired,
-  item: PropTypes.shape({
-    LinkWrapper: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-    path: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-  }).isRequired,
-};
-
 const TopLevelMenuToggle = styled(Link).attrs({ isButton: true, tertiary: true })`
   font-weight: ${typography.weight.bold};
-`;
-
-const MenuLink = styled(({ isActive, ...rest }) => <Link {...rest} />)`
-  color: ${(props) => (props.isActive ? color.secondary : color.darkest)};
-  font-weight: ${(props) => (props.isActive ? typography.weight.bold : typography.weight.regular)};
 `;
 
 const MenuChild = styled.div`
@@ -186,36 +97,6 @@ Menu.propTypes = {
 
 Menu.defaultProps = {
   isTopLevel: false,
-};
-
-const ItemLinkWrapper = styled.li`
-  &:last-of-type ${MenuLink} {
-    margin-bottom: 0;
-  }
-`;
-
-function ItemLink({ currentPath, item }) {
-  return (
-    <ItemLinkWrapper>
-      <MenuLink
-        isActive={currentPath === item.path}
-        href={item.path}
-        LinkWrapper={item.LinkWrapper}
-        tertiary
-      >
-        {item.title}
-      </MenuLink>
-    </ItemLinkWrapper>
-  );
-}
-
-ItemLink.propTypes = {
-  currentPath: PropTypes.string.isRequired,
-  item: PropTypes.shape({
-    path: PropTypes.string.isRequired,
-    LinkWrapper: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-    title: PropTypes.string.isRequired,
-  }).isRequired,
 };
 
 const List = styled.ul`
